refactor: restructure to full atomic-design architecture
