--- src/hooks/render.ts
@@ -1,4 +1,0 @@
-import { RenderContext } from "@/context/render";
-import React from "react";
-
-export const useRenderContext = () => React.useContext(RenderContext);
