+++ src/ions/constants/index.ts
@@ -1,0 +1,1 @@
+export const ROOT = "ROOT";
