+++ src/ions/contexts/render.ts
@@ -1,0 +1,4 @@
+import { Renderer } from "@/ions/enums";
+import React from "react";
+
+export const RenderContext = React.createContext({ renderer: Renderer.html });
