+++ src/ions/hooks/render.ts
@@ -1,0 +1,4 @@
+import { RenderContext } from "@/ions/contexts/render";
+import React from "react";
+
+export const useRenderContext = () => React.useContext(RenderContext);
