+++ src/molecules/language-switcher/index.tsx
@@ -1,0 +1,49 @@
+import { StyledLink } from "@/atoms/active-link";
+import { ErrorComponent } from "@/atoms/error-component";
+import { ROOT } from "@/ions/constants";
+import routes from "@/ions/routes";
+import { LocalizedString } from "@/types";
+import { useTranslation } from "next-i18next";
+import Link from "next/link";
+import { useRouter } from "next/router";
+import React from "react";
+
+export const LanguageSwitcher: React.FC = () => {
+	const router = useRouter();
+	const { t } = useTranslation("common");
+	const [localizedDir] = router.query.args ?? [ROOT];
+	const found = Object.entries(routes).find(([, value]) => {
+		return Object.values(value.dir).includes(localizedDir);
+	});
+
+	if (!found && localizedDir !== ROOT) {
+		return <ErrorComponent message={`Could not find href for ${localizedDir}`} />;
+	}
+
+	const [, { dir }] = found ?? [
+		ROOT,
+		{
+			dir: router.locales.reduce(
+				(previousValue, locale) => ({ ...previousValue, [locale]: "" }),
+				{}
+			) as LocalizedString,
+		},
+	];
+
+	return (
+		<div>
+			{router.locales.map(locale => (
+				<div key={locale}>
+					<Link passHref href={`/${dir[locale]}`} locale={locale}>
+						<StyledLink
+							active={router.locale === locale}
+							data-test-id="language-switcher-link"
+						>
+							{t(locale)}
+						</StyledLink>
+					</Link>
+				</div>
+			))}
+		</div>
+	);
+};
