+++ src/organisms/footer/styled.ts
@@ -1,0 +1,15 @@
+import styled from "@emotion/styled";
+
+export const StyledFooter = styled.footer`
+	position: absolute;
+	z-index: 10;
+	right: 0;
+	bottom: 0;
+	left: 0;
+	width: 100%;
+	max-width: 1280px;
+	margin: 0 auto;
+	padding: 2em;
+	background: white;
+	color: black;
+`;
