+++ src/organisms/header/styled.ts
@@ -1,0 +1,15 @@
+import styled from "@emotion/styled";
+
+export const StyledHeader = styled.header`
+	position: absolute;
+	z-index: 10;
+	top: 0;
+	right: 0;
+	left: 0;
+	width: 100%;
+	max-width: 1280px;
+	margin: 0 auto;
+	padding: 2em;
+	background: white;
+	color: black;
+`;
