+++ src/templates/home/index.tsx
@@ -1,0 +1,6 @@
+import { Layout } from "@/organisms/layout";
+import React from "react";
+
+export const Template: React.FC = ({ children }) => {
+	return <Layout>{children}</Layout>;
+};
